requirements.txt
aiogram==3.13.1
mistralai==1.2.5
python-dotenv==1.0.1
Pillow==10.4.0
aiohttp==3.10.5
faster-whisper==1.0.3
PyMuPDF==1.24.11
python-docx==1.1.2
httpx==0.27.2
pydantic==2.9.2
aiofiles==24.1.0
